requirements.txt
django
psycopg2-binary
requests
numpy
